Refactor EvoEngine to explicit application-owned runtime state with install deployment, runtime package support, refreshed docs, LF formatting, and SDK warning cleanup.
refactor: move runtime ownership into Application
feat: add runtime package loading infrastructure
build: add install deployment and simplify VS presets
docs: refresh README and plugin documentation
chore: add LF normalization and C++ formatting helper
fix: clean SDK narrowing warnings

+++ EvoEngine_Plugins/EcoSysLab/src/IVolume.cpp
@@ -15,14 +15,14 @@ bool IVolume::InVolume(const glm::vec3& position) {
 void IVolume::InVolume(const GlobalTransform& global_transform, const std::vector<glm::vec3>& positions,
                        std::vector<bool>& results) {
   results.resize(positions.size());
-  Jobs::RunParallelFor(positions.size(), [&](unsigned i) {
+  Jobs::RunParallelFor(positions.size(), [&](size_t i) {
     results[i] = InVolume(global_transform, positions[i]);
   });
 }
 
 void IVolume::InVolume(const std::vector<glm::vec3>& positions, std::vector<bool>& results) {
   results.resize(positions.size());
-  Jobs::RunParallelFor(positions.size(), [&](unsigned i) {
+  Jobs::RunParallelFor(positions.size(), [&](size_t i) {
     results[i] = InVolume(positions[i]);
   });
 }
